Expand HybridLatestAt functionality when it's not being used via range
