+++ crates/re_space_view/src/lib.rs
@@ -11,8 +11,8 @@ mod screenshot;
 mod view_property_ui;
 
 pub use heuristics::suggest_space_view_for_each_entity;
-pub use query::{latest_at_with_overrides, range_with_overrides, HybridResults};
-pub use results_ext::RangeResultsExt;
+pub use query::{latest_at_with_overrides, range_with_overrides};
+pub use results_ext::{HybridResults, RangeResultsExt};
 pub use screenshot::ScreenshotMode;
 pub use view_property_ui::view_property_ui;
